@@ -439,9 +439,9 @@ class DisplayGroup extends Base
      *      required=true
      *   ),
      *  @SWG\Parameter(
-     *      name="dynamicContent",
+     *      name="dynamicCriteria",
      *      in="formData",
-     *      description="The filter criteria for this dynamic group. A command separated set of regular expressions to apply",
+     *      description="The filter criteria for this dynamic group. A comma separated set of regular expressions to apply",
      *      type="string",
      *      required=false
      *   ),
