@@ -63,6 +63,32 @@ class Ticker extends ModuleWidget
     }
 
     /**
+     * Form for updating the module settings
+     */
+    public function settingsForm()
+    {
+        return 'ticker-form-settings';
+    }
+
+    /**
+     * Process any module settings
+     * @throws InvalidArgumentException
+     */
+    public function settings()
+    {
+        $updateIntervalImages = $this->getSanitizer()->getInt('updateIntervalImages', 240);
+
+        if ($this->module->enabled != 0) {
+            if ($updateIntervalImages < 0)
+                throw new InvalidArgumentException(__('Update Interval Images must be greater than or equal to 0'), 'updateIntervalImages');
+        }
+
+        $this->module->settings['updateIntervalImages'] = $updateIntervalImages;
+
+        return $this->module->settings;
+    }
+
+    /**
      * Get Extra content for the form
      * @return array
      */
@@ -121,6 +147,13 @@ class Ticker extends ModuleWidget
      *      name="updateInterval",
      *      in="formData",
      *      description="Update interval in minutes",
+     *      type="integer",
+     *      required=false
+     *   ),
+     *  @SWG\Parameter(
+     *      name="updateIntervalImages",
+     *      in="formData",
+     *      description="Update interval for downloaded Images, in minutes",
      *      type="integer",
      *      required=false
      *   ),
@@ -281,6 +314,11 @@ class Ticker extends ModuleWidget
         $this->setOption('xmds', true);
         $this->setOption('uri', urlencode($this->getSanitizer()->getString('uri')));
         $this->setOption('updateInterval', $this->getSanitizer()->getInt('updateInterval', 120));
+
+        if ($this->getSanitizer()->getInt('updateIntervalImages') !== null) {
+            $this->setOption('updateIntervalImages', $this->getSanitizer()->getInt('updateIntervalImages'));
+        }
+
         $this->setOption('speed', $this->getSanitizer()->getInt('speed', 2));
         $this->setOption('name', $this->getSanitizer()->getString('name'));
         $this->setOption('effect', $this->getSanitizer()->getString('effect'));
@@ -649,7 +687,7 @@ class Ticker extends ModuleWidget
         $dateFormat = $this->getOption('dateFormat', $this->getConfig()->getSetting('DATE_FORMAT'));
 
         // Set an expiry time for the media
-        $expires = $this->getDate()->parse()->addMinutes($this->getOption('updateInterval', 3600))->format('U');
+        $expiresImage = $this->getDate()->parse()->addMinutes($this->getOption('updateIntervalImages', $this->getSetting('updateIntervalImages', 1440)))->format('U');
 
         // Render the content now
         foreach ($feedItems as $item) {
@@ -722,7 +760,7 @@ class Ticker extends ModuleWidget
                         // image url
                         if ($link != NULL) {
                             // Grab the profile image
-                            $file = $this->mediaFactory->queueDownload('ticker_' . md5($this->getOption('url') . $link), $link, $expires);
+                            $file = $this->mediaFactory->queueDownload('ticker_' . md5($this->getOption('url') . $link), $link, $expiresImage);
 
                             $replace = '<img src="' . $this->getFileUrl($file, 'image') . '" ' . $attribute . ' />';
                         }
@@ -789,7 +827,7 @@ class Ticker extends ModuleWidget
 
                                 if (!(empty($link))) {
                                     // Grab the image
-                                    $file = $this->mediaFactory->queueDownload('ticker_' . md5($this->getOption('url') . $link), $link, $expires);
+                                    $file = $this->mediaFactory->queueDownload('ticker_' . md5($this->getOption('url') . $link), $link, $expiresImage);
 
                                     $replace = '<img src="' . $this->getFileUrl($file, 'image') . '"/>';
                                 } else {
@@ -854,8 +892,11 @@ class Ticker extends ModuleWidget
         if (!v::numeric()->validate($this->getOption('numItems', 0)))
             throw new InvalidArgumentException(__('The value in Number of Items must be numeric.'), 'numItems');
 
-        if (!v::intType()->min(0)->validate($this->getOption('updateInterval')))
+        if ($this->getOption('updateInterval') !== null && !v::intType()->min(0)->validate($this->getOption('updateInterval', 0)))
             throw new InvalidArgumentException(__('Update Interval must be greater than or equal to 0'), 'updateInterval');
+
+        if ($this->getOption('updateIntervalImages') !== null && !v::intType()->min(0)->validate($this->getOption('updateIntervalImages', 0)))
+            throw new InvalidArgumentException(__('Update Interval Images must be greater than or equal to 0'), 'updateIntervalImages');
 
         return self::$STATUS_VALID;
     }
@@ -871,6 +912,12 @@ class Ticker extends ModuleWidget
     {
         // Tickers are non-display specific
         return $this->getWidgetId() . (($displayId === 0) ? '_0' : '');
+    }
+
+    /** @inheritdoc */
+    public function isCacheDisplaySpecific()
+    {
+        return ($this->getOption('sourceId', 1) == 2);
     }
 
     /** @inheritdoc */
