@@ -153,9 +153,13 @@ class Soap4 extends Soap
                 $display->auditingUntil = 0;
                 $display->defaultLayoutId = $this->getConfig()->getSetting('DEFAULT_LAYOUT', 4);
                 $display->license = $hardwareKey;
-                $display->licensed = 0;
+                $display->licensed = $this->getConfig()->getSetting('DISPLAY_AUTO_AUTH', 0);
                 $display->incSchedule = 0;
                 $display->clientAddress = $this->getIp();
+
+                if (!$display->maxLicensedDisplays()) {
+                    $display->licensed = 0;
+                }
             }
             catch (\InvalidArgumentException $e) {
                 throw new \SoapFault('Sender', $e->getMessage());
@@ -163,7 +167,10 @@ class Soap4 extends Soap
 
             $displayElement->setAttribute('status', 1);
             $displayElement->setAttribute('code', 'ADDED');
-            $displayElement->setAttribute('message', 'Display added and is awaiting licensing approval from an Administrator.');
+            if ($display->licensed == 0)
+                $displayElement->setAttribute('message', 'Display added and is awaiting licensing approval from an Administrator.');
+            else
+                $displayElement->setAttribute('message', 'Display is active and ready to start.');
         }
 
 
