@@ -7,9 +7,11 @@
 
 namespace Xibo\Tests;
 
+use Monolog\Handler\NullHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
 use Slim\Environment;
 use Slim\Helper\Set;
 use Slim\Log;
@@ -82,11 +84,16 @@ class LocalWebTestCase extends WebTestCase
         ]);
 
         // Create a logger
+        $handlers = [];
+        if (isset($_SERVER['PHPUNIT_LOG_TO_FILE']) && $_SERVER['PHPUNIT_LOG_TO_FILE']) {
+            $handlers[] = new StreamHandler(PROJECT_ROOT . '/library/log.txt', Logger::DEBUG);
+        } else {
+            $handlers[] = new NullHandler();
+        }
+
         $logger = new AccessibleMonologWriter(array(
             'name' => 'PHPUNIT',
-            'handlers' => array(
-                new StreamHandler(PROJECT_ROOT . '/library/log.txt', Logger::DEBUG)
-            ),
+            'handlers' => $handlers,
             'processors' => array(
                 new \Xibo\Helper\LogProcessor(),
                 new \Monolog\Processor\UidProcessor(7)
@@ -298,8 +305,11 @@ class LocalWebTestCase extends WebTestCase
     {
         // Create if necessary
         if (self::$logger === null) {
-            self::$logger = new Logger('TESTS', [new \Monolog\Handler\StreamHandler(STDERR, Logger::DEBUG)]);
-            //self::$logger = new NullLogger();
+            if (isset($_SERVER['PHPUNIT_LOG_TO_CONSOLE']) && $_SERVER['PHPUNIT_LOG_TO_CONSOLE']) {
+                self::$logger = new Logger('TESTS', [new \Monolog\Handler\StreamHandler(STDERR, Logger::DEBUG)]);
+            } else {
+                self::$logger = new NullLogger();
+            }
         }
 
         return self::$logger;
@@ -316,5 +326,39 @@ class LocalWebTestCase extends WebTestCase
             $this->fail('Test hasnt used the client and therefore cannot determine XMR activity');
 
         return $service->processQueue();
+    }
+
+    protected static function installModuleIfNecessary($name, $class)
+    {
+        // Make sure the HLS widget is installed
+        $res = self::$container->store->select('SELECT * FROM `module` WHERE `module` = :module', ['module' => $name]);
+
+        if (count($res) <= 0) {
+            // Install the module
+            self::$container->store->insert('
+              INSERT INTO `module` (`Module`, `Name`, `Enabled`, `RegionSpecific`, `Description`,
+                `ImageUri`, `SchemaVersion`, `ValidExtensions`, `PreviewEnabled`, `assignable`, `render_as`, `settings`, `viewPath`, `class`, `defaultDuration`, `installName`)
+              VALUES (:module, :name, :enabled, :region_specific, :description,
+                :image_uri, :schema_version, :valid_extensions, :preview_enabled, :assignable, :render_as, :settings, :viewPath, :class, :defaultDuration, :installName)
+            ', [
+                'module' => $name,
+                'name' => $name,
+                'enabled' => 1,
+                'region_specific' => 1,
+                'description' => $name,
+                'image_uri' => null,
+                'schema_version' => 1,
+                'valid_extensions' => null,
+                'preview_enabled' => 1,
+                'assignable' => 1,
+                'render_as' => 'html',
+                'settings' => json_encode([]),
+                'viewPath' => '../modules',
+                'class' => $class,
+                'defaultDuration' => 10,
+                'installName' => $name
+            ]);
+            self::$container->store->commitIfNecessary();
+        }
     }
 }
