@@ -73,11 +73,10 @@ class DataSetDataEditTest extends LocalWebTestCase
         $layout = $this->checkout($this->layout);
 
         // Add a couple of text widgets to the region
-        $response = $this->getEntityProvider()->post('/playlist/widget/datasetview/' . $layout->regions[0]->regionPlaylist['playlistId'], [
-            'name' => Random::generateString(),
-            'dataSetId' => $this->dataSet->dataSetId,
-            'duration' => 100,
-            'useDuration' => 1
+        $response = $this->getEntityProvider()->post('/playlist/widget/datasetview/' . $layout->regions[0]->regionPlaylist['playlistId']);
+        $response = $this->getEntityProvider()->put('/playlist/widget/' . $response['widgetId'], [
+            'step' => 1,
+            'dataSetId' => $this->dataSet->dataSetId
         ]);
 
         $this->widget = (new XiboDataSetView($this->getEntityProvider()))->hydrate($response);
