@@ -53,22 +53,18 @@ class GetResourceTest extends LocalWebTestCase
         $layout = $this->checkout($this->layout);
 
         // Add a resource heavy module to the Layout (one that will download images)
-        $response = $this->getEntityProvider()->post('/playlist/widget/ticker/' . $layout->regions[0]->regionPlaylist['playlistId'], [
+        $response = $this->getEntityProvider()->post('/playlist/widget/ticker/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $response = $this->getEntityProvider()->put('/playlist/widget/' . $response['widgetId'], [
             'uri' => 'http://ceu.xibo.co.uk/mediarss/feed.xml',
             'duration' => 100,
             'useDuration' => 1,
             'sourceId' => 1,
+            'templateId' => 'media-rss-with-title'
         ]);
 
         // Edit the Ticker to add the template
         $this->widget = (new XiboTicker($this->getEntityProvider()))->hydrate($response);
-
-        $this->getEntityProvider()->put('/playlist/widget/' . $this->widget->widgetId, [
-            'uri' => 'http://ceu.xibo.co.uk/mediarss/feed.xml',
-            'duration' => 100,
-            'useDuration' => 1,
-            'templateId' => 'media-rss-with-title'
-        ]);
 
         // Checkin
         $this->layout = $this->publish($this->layout);
