@@ -1,7 +1,8 @@
 <?php
-/*
+/**
+ * Copyright (C) 2018 Xibo Signage Ltd
+ *
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2015-2018 Spring Signage Ltd
  *
  * This file is part of Xibo.
  *
@@ -22,101 +23,56 @@
 namespace Xibo\Tests\Integration\Widget;
 
 use Xibo\OAuth2\Client\Entity\XiboClock;
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
 
+/**
+ * Class ClockWidgetTest
+ * @package Xibo\Tests\Integration\Widget
+ */
 class ClockWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/clock/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
+
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
-
-	/**
-	 * @group add
-     * @dataProvider provideSuccessCases
-     */
-	public function testAdd($name, $duration, $useDuration, $theme, $clockTypeId, $offset, $format, $showSeconds, $clockFace)
-	{
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-		$response = $this->client->post('/playlist/widget/clock/' . $playlistId, [
-        	'name' => $name,
-        	'duration' => $duration,
-            'useDuration' => $useDuration,
-        	'themeId' => $theme,
-        	'clockTypeId' => $clockTypeId,
-        	'offset' => $offset,
-        	'format' => $format,
-        	'showSeconds' => $showSeconds,
-        	'clockFace' => $clockFace
-        	]);
-        $this->assertSame(200, $this->client->response->status());
-        $this->assertNotEmpty($this->client->response->body());
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $clockOptions = (new XiboClock($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($name, $clockOptions->name);
-        $this->assertSame($duration, $clockOptions->duration);
-        
-        foreach ($clockOptions->widgetOptions as $option) {
-            if ($option['option'] == 'theme') {
-                $this->assertSame($theme, intval($option['value']));
-            }
-            if ($option['option'] == 'clockTypeId') {
-                $this->assertSame($clockTypeId, intval($option['value']));
-            }
-            if ($option['option'] == 'offset') {
-                $this->assertSame($offset, intval($option['value']));
-            }
-            if ($option['option'] == 'format') {
-                $this->assertSame($format, $option['value']);
-            }
-            if ($option['option'] == 'showSeconds') {
-                $this->assertSame($showSeconds, intval($option['value']));
-            }
-            if ($option['option'] == 'clockFace') {
-                $this->assertSame($clockFace, $option['value']);
-            }
-        }
-	}
 
 	/**
      * Each array is a test run
@@ -134,53 +90,38 @@ class ClockWidgetTest extends LocalWebTestCase
         ];
     }
 
-    public function testEdit()
+    /**
+     * @throws \Xibo\OAuth2\Client\Exception\XiboApiException
+     * @dataProvider provideSuccessCases
+     */
+    public function testEdit($name, $duration, $useDuration, $theme, $clockTypeId, $offset, $format, $showSeconds, $clockFace)
     {
-        # Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-    	# Create a clock with wrapper
-    	$clock = (new XiboClock($this->getEntityProvider()))->create('Api Analogue clock', 20, 1, 1, 1, NULL, NULL, NULL, NULL, $playlistId);
-    	$nameNew = 'Edited Name';
-    	$durationNew = 80;
-    	$clockTypeIdNew = 3;
-    	$response = $this->client->put('/playlist/widget/' . $clock->widgetId, [
-        	'name' => $nameNew,
-        	'duration' => $durationNew,
-        	'themeId' => $clock->theme,
-        	'clockTypeId' => $clockTypeIdNew,
-        	'offset' => $clock->offset,
-        	'format' => $clock->format,
-        	'showSeconds' => $clock->showSeconds,
-        	'clockFace' => $clock->clockFace
+        $response = $this->client->put('/playlist/widget/' . $this->widgetId, [
+        	'name' => $name,
+        	'duration' => $duration,
+        	'themeId' => $theme,
+        	'clockTypeId' => $clockTypeId,
+        	'offset' => $offset,
+        	'format' => $format,
+        	'showSeconds' => $showSeconds,
+        	'clockFace' => $clockFace
         	], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+
         $this->assertSame(200, $this->client->response->status());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $clockOptions = (new XiboClock($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($nameNew, $clockOptions->name);
-        $this->assertSame($durationNew, $clockOptions->duration);
-        foreach ($clockOptions->widgetOptions as $option) {
+
+        /** @var XiboClock $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboClock($this->getEntityProvider()))->hydrate($response[0]);
+
+        foreach ($checkWidget->widgetOptions as $option) {
             if ($option['option'] == 'clockTypeId') {
-                $this->assertSame($clockTypeIdNew, intval($option['value']));
+                $this->assertSame($clockTypeId, intval($option['value']));
+            } else if ($option['option'] == 'name') {
+                $this->assertSame($name, $option['value']);
             }
         }
-    }
-    public function testDelete()
-    {
-        # Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-    	# Create a clock with wrapper
-		$clock = (new XiboClock($this->getEntityProvider()))->create('Api Analogue clock', 20, 1, 1, 1, NULL, NULL, NULL, NULL, $playlistId);
-		# Delete it
-		$this->client->delete('/playlist/widget/' . $clock->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
     }
 }
