@@ -1,7 +1,8 @@
 <?php
-/*
+/**
+ * Copyright (C) 2018 Xibo Signage Ltd
+ *
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2015-2018 Spring Signage Ltd
  *
  * This file is part of Xibo.
  *
@@ -22,131 +23,58 @@
 namespace Xibo\Tests\Integration\Widget;
 
 use Xibo\OAuth2\Client\Entity\XiboCurrencies;
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
 
+/**
+ * Class CurrenciesWidgetTest
+ * @package Xibo\Tests\Integration\Widget
+ */
 class CurrenciesWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/currencies/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
+
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
 
-    /**
-     * @dataProvider provideSuccessCases
-     * @group broken
-     */
-    public function testAdd($isOverride, $templateId, $name, $duration, $useDuration, $base, $items, $reverseConversion, $effect, $speed, $backgroundColor, $noRecordsMessage, $dateFormat, $updateInterval, $durationIsPerPage, $widgetOriginalWidth, $widgetOriginalHeight, $maxItemsPerPage, $mainTemplate, $itemTemplate, $styleSheet, $javaScript)
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-        
-        if ($isOverride) {
-            $response = $this->client->post('/playlist/widget/currencies/' . $playlistId, [
-                'templateId' => $templateId,
-                'name' => $name,
-                'duration' => $duration,
-                'useDuration' => $useDuration,
-                'base' => $base,
-                'items' => $items,
-                'reverseConversion' => $reverseConversion,
-                'effect' => $effect,
-                'speed' => $speed,
-                'backgroundColor' => $backgroundColor,
-                'noRecordsMessage' => $noRecordsMessage,
-                'dateFormat' => $dateFormat,
-                'updateInterval' => $updateInterval,
-                'overrideTemplate' => 1,
-                'durationIsPerPage' => $durationIsPerPage,
-                'widgetOriginalWidth' => $widgetOriginalWidth,
-                'widgetOriginalHeight' => $widgetOriginalHeight,
-                'maxItemsPerPage' => $maxItemsPerPage,
-                'mainTemplate' => $mainTemplate,
-                'itemTemplate' => $itemTemplate,
-                'styleSheet' => $styleSheet,
-                'javaScript' => $javaScript
-            ]);
-        } else {
-            $response = $this->client->post('/playlist/widget/currencies/' . $playlistId, [
-                'templateId' => $templateId,
-                'name' => $name,
-                'duration' => $duration,
-                'useDuration' => $useDuration,
-                'base' => $base,
-                'items' => $items,
-                'reverseConversion' => $reverseConversion,
-                'effect' => $effect,
-                'speed' => $speed,
-                'backgroundColor' => $backgroundColor,
-                'noRecordsMessage' => $noRecordsMessage,
-                'dateFormat' => $dateFormat,
-                'updateInterval' => $updateInterval,
-                'overrideTemplate' => 0,
-                'durationIsPerPage' => $durationIsPerPage,
-            ]);
-        }
 
-        $widgetOptions = (new XiboCurrencies($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($name, $widgetOptions->name);
-        $this->assertSame($duration, $widgetOptions->duration);
-        foreach ($widgetOptions->widgetOptions as $option) {
-            if ($option['option'] == 'templateId') {
-                $this->assertSame($templateId, $option['value']);
-            }
-            if ($option['option'] == 'base') {
-                $this->assertSame($base, $option['value']);
-            }
-            if ($option['option'] == 'items') {
-                $this->assertSame($items, $option['value']);
-            }
-            if ($option['option'] == 'updateInterval') {
-                $this->assertSame($updateInterval, intval($option['value']));
-            }
-            if ($option['option'] == 'reverseConversion') {
-                $this->assertSame($reverseConversion, intval($option['value']));
-            }
-            if ($option['option'] == 'maxItemsPerPage') {
-                $this->assertSame($maxItemsPerPage, intval($option['value']));
-            }
-        }
-
-    }
 
     /**
      * Each array is a test run
@@ -165,74 +93,51 @@ class CurrenciesWidgetTest extends LocalWebTestCase
 
     /**
      * @group broken
+     * @dataProvider provideSuccessCases
      */
-    public function testEdit()
+    public function testEdit($isOverride, $templateId, $name, $duration, $useDuration, $base, $items, $reverseConversion, $effect, $speed, $backgroundColor, $noRecordsMessage, $dateFormat, $updateInterval, $durationIsPerPage, $widgetOriginalWidth, $widgetOriginalHeight, $maxItemsPerPage, $mainTemplate, $itemTemplate, $styleSheet, $javaScript)
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a currencies with wrapper
-        $currencies = (new XiboCurrencies($this->getEntityProvider()))->create('currencies2', 'Unedited widget', 120, 1, 'GBP', 'EUR', 1, NULL, NULL, NULL, 'No messages', NULL, 50, 1, $playlistId);
-        $nameNew = 'Edited widget';
-        $durationNew = 80;
-        $templateNew = 'currencies1';
-        $notReverse = 0;
-        $itemsNew = 'USD';
         # Edit currency widget and change name, duration, template, reverseConversion and items
-        $response = $this->client->put('/playlist/widget/' . $currencies->widgetId, [
-                'templateId' => $templateNew,
-                'name' => $nameNew,
-                'duration' => $durationNew,
-                'useDuration' => $currencies->useDuration,
-                'base' => $currencies->base,
-                'items' => $itemsNew,
-                'reverseConversion' => $notReverse,
-                'effect' => $currencies->effect,
-                'speed' => $currencies->speed,
-                'backgroundColor' => $currencies->backgroundColor,
-                'noRecordsMessage' => $currencies->noRecordsMessage,
-                'dateFormat' => $currencies->dateFormat,
-                'updateInterval' => $currencies->updateInterval,
-                'durationIsPerPage' => $currencies->durationIsPerPage,
+        $response = $this->client->put('/playlist/widget/' . $this->widgetId, [
+                'templateId' => $templateId,
+                'name' => $name,
+                'duration' => $duration,
+                'useDuration' => $useDuration,
+                'base' => $base,
+                'items' => $items,
+                'reverseConversion' => $reverseConversion,
+                'effect' => $effect,
+                'speed' => $speed,
+                'backgroundColor' => $backgroundColor,
+                'noRecordsMessage' => $noRecordsMessage,
+                'dateFormat' => $dateFormat,
+                'updateInterval' => $updateInterval,
+                'durationIsPerPage' => $durationIsPerPage,
         ], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+
         $this->assertSame(200, $this->client->response->status());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $widgetOptions = (new XiboCurrencies($this->getEntityProvider()))->getById($playlistId);
+
+        /** @var XiboCurrencies $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboCurrencies($this->getEntityProvider()))->hydrate($response[0]);
+
         # check if changes were correctly saved
-        $this->assertSame($nameNew, $widgetOptions->name);
-        $this->assertSame($durationNew, $widgetOptions->duration);
-        foreach ($widgetOptions->widgetOptions as $option) {
+        $this->assertSame($name, $checkWidget->name);
+        $this->assertSame($duration, $checkWidget->duration);
+
+        foreach ($checkWidget->widgetOptions as $option) {
             if ($option['option'] == 'templateId') {
-                $this->assertSame($templateNew, $option['value']);
+                $this->assertSame($templateId, $option['value']);
             }
             if ($option['option'] == 'items') {
-                $this->assertSame($itemsNew, $option['value']);
+                $this->assertSame($items, $option['value']);
             }
             if ($option['option'] == 'reverseConversion') {
-                $this->assertSame($notReverse, intval($option['value']));
+                $this->assertSame($reverseConversion, intval($option['value']));
             }
         }
-    }
-
-    /**
-     * @group broken
-     */
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a currencies with wrapper
-        $currencies = (new XiboCurrencies($this->getEntityProvider()))->create('currencies2', 'Unedited widget', 120, 1, 'GBP', 'EUR', 1, NULL, NULL, NULL, 'No messages', NULL, 50, 1, $playlistId);
-        # Delete it
-        $this->client->delete('/playlist/widget/' . $currencies->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
     }
 }
