@@ -1,7 +1,8 @@
 <?php
-/*
+/**
+ * Copyright (C) 2018 Xibo Signage Ltd
+ *
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2015-2018 Spring Signage Ltd
  *
  * This file is part of Xibo.
  *
@@ -21,88 +22,69 @@
 
 namespace Xibo\Tests\Integration\Widget;
 
+use Xibo\Helper\Random;
 use Xibo\OAuth2\Client\Entity\XiboEmbedded;
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
 
+/**
+ * Class EmbeddedWidgetTest
+ * @package Xibo\Tests\Integration\Widget
+ */
 class EmbeddedWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/embedded/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
+
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
 
-    public function testAdd()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        $response = $this->client->post('/playlist/widget/embedded/' . $playlistId, [
-            'name' => 'API Embedded widget',
-            'duration' => 60,
-            'transparency' => 0,
-            'scaleContent' => 0,
-            'embedHtml' => null,
-            'embedScript' => null,
-            'embedStyle' => null
-            ]);
-
-        $this->assertSame(200, $this->client->response->status(), "Not successful: " . $response);
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object);
-        $embeddedOptions = (new XiboEmbedded($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame('API Embedded widget', $embeddedOptions->name);       
-        $this->assertSame(60, $embeddedOptions->duration);
-    }
-
+    /**
+     * @throws \Xibo\OAuth2\Client\Exception\XiboApiException
+     */
     public function testEdit()
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
+        $name = Random::generateString();
         $durationNew = 80;
-        # Create embedded widget
-        $embedded = (new XiboEmbedded($this->getEntityProvider()))->create('API embedded', 60, 1, 0, 0, null, null, null, $playlistId);
-        $response = $this->client->put('/playlist/widget/' . $embedded->widgetId, [
-            'name' => 'EDITED Name',
+
+        $response = $this->client->put('/playlist/widget/' . $this->widgetId, [
+            'name' => $name,
             'duration' => $durationNew,
             'transparency' => 1,
             'scaleContent' => 1,
@@ -110,27 +92,17 @@ class EmbeddedWidgetTest extends LocalWebTestCase
             'embedScript' => null,
             'embedStyle' => '<style type="text/css"> </style>'
             ], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+
         $this->assertSame(200, $this->client->response->status());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $embeddedOptions = (new XiboEmbedded($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame('EDITED Name', $embeddedOptions->name);       
-        $this->assertSame($durationNew, $embeddedOptions->duration);
-    }
 
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
+        /** @var XiboEmbedded $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboEmbedded($this->getEntityProvider()))->hydrate($response[0]);
 
-        # Create embedded widget
-        $embedded = (new XiboEmbedded($this->getEntityProvider()))->create('API embedded', 60, 1, 0, 0, null, null, null, $playlistId);
-        # Delete it
-        $this->client->delete('/playlist/widget/' . $embedded->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
+        $this->assertSame($name, $checkWidget->name);
+        $this->assertSame($durationNew, $checkWidget->duration);
     }
 }
