@@ -1,7 +1,8 @@
 <?php
-/*
+/**
+ * Copyright (C) 2018 Xibo Signage Ltd
+ *
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2015-2018 Spring Signage Ltd
  *
  * This file is part of Xibo.
  *
@@ -22,192 +23,117 @@
 namespace Xibo\Tests\Integration\Widget;
 
 use Xibo\OAuth2\Client\Entity\XiboHls;
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
 
+/**
+ * Class HlsWidgetTest
+ * @package Xibo\Tests\Integration\Widget
+ */
 class HlsWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
+    public static function setUpBeforeClass()
+    {
+        parent::setUpBeforeClass();
+
+        parent::installModuleIfNecessary('hls', '\Xibo\Widget\Hls');
+    }
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/hls/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
+
+        $this->getLogger()->debug('Setup Finished');
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
+
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
-
-	/**
-     * @group broken
-     * @dataProvider provideSuccessCases
-     */
-	public function testAdd($name, $useDuration, $duration, $uri, $mute, $transparency)
-	{
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-		$response = $this->client->post('/playlist/widget/hls/' . $playlistId, [
-        	'name' => $name,
-            'useDuration' => $useDuration,
-        	'duration' => $duration,
-        	'uri' => $uri,
-        	'mute' => $mute,
-        	'transparency' => $transparency,
-        	]);
-        $this->assertSame(200, $this->client->response->status());
-        $this->assertNotEmpty($this->client->response->body());
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $widgetOptions = (new XiboHls($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($name, $widgetOptions->name);
-        $this->assertSame($duration, $widgetOptions->duration);
-        
-        foreach ($widgetOptions->widgetOptions as $option) {
-            if ($option['option'] == 'uri') {
-                $this->assertSame($uri, urldecode(($option['value'])));
-            }
-        }
-	}
 
 	/**
      * Each array is a test run
      * Format ($name, $useDuration, $duration, $uri, $mute, $transparency)
      * @return array
      */
-    public function provideSuccessCases()
+    public function provideEditCases()
     {
         # Sets of data used in testAdd
         return [
-            'HLS stream' => ['HLS stream', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_master.m3u8', 0, 0],
-            'HLS stream 512' => ['HLS stream with transparency', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_512.m3u8', 0, 1],
+            'HLS stream' => [200, 'HLS stream', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_master.m3u8', 0, 0],
+            'HLS stream 512' => [200, 'HLS stream with transparency', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_512.m3u8', 0, 1],
+            'No url provided' => [500, 'no uri', 1, 10, '', 0, 0],
+            'No duration provided' => [500, 'no duration with useDuration 1', 1, 0, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_512.m3u8', 0, 0],
         ];
     }
 
     /**
-     * testAddFailure - test adding various hls widgets that should be invalid
-     * @dataProvider provideFailureCases
-     * @group broken
-     */
-    public function testAddFailure($name, $useDuration, $duration, $uri, $mute, $transparency)
+    * Edit
+    * @dataProvider provideEditCases
+    */
+    public function testEdit($statusCode, $name, $useDuration, $duration, $uri, $mute, $transparency)
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create Google traffic widgets with arguments from provideFailureCases
-        $response = $this->client->post('/playlist/widget/hls/' . $playlistId, [
+        $response = $this->client->put('/playlist/widget/' . $this->widgetId, [
             'name' => $name,
             'useDuration' => $useDuration,
             'duration' => $duration,
             'uri' => $uri,
             'mute' => $mute,
             'transparency' => $transparency,
-            ]);
-        # check if they fail as expected
-        $this->assertSame(500, $this->client->response->status(), 'Expecting failure, received ' . $this->client->response->status());
-    }
-
-    /**
-     * Each array is a test run
-     * Format ($$name, $useDuration, $duration, $uri, $mute, $transparency)
-     * @return array
-     */
-    public function provideFailureCases()
-    {
-        # Data for testAddfailure, easily expandable - just add another set of data below
-        return [
-            'No url provided' => ['no uri', 1, 10, null, 0, 0],
-            'No duration provided' => ['no duration with useDuration 1', 1, 0, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_512.m3u8', 0, 0],
-        ];
-    }
-
-    /**
-    * Edit
-    * @group broken
-    */
-    public function testEdit()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create Google traffic widget
-        $hls = (new XiboHls($this->getEntityProvider()))->create('HLS stream', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_master.m3u8', 0, 0, $playlistId);
-        $nameNew = 'Edited Widget';
-        $durationNew = 100;
-        $uriNew = 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_512.m3u8';
-        $response = $this->client->put('/playlist/widget/' . $hls->widgetId, [
-            'name' => $nameNew,
-            'useDuration' => 1,
-            'duration' => $durationNew,
-            'uri' => $uriNew,
-            'mute' => $hls->mute,
-            'transparency' => $hls->transparency,
             ], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
-        $this->assertSame(200, $this->client->response->status());
+
+        $this->assertSame($statusCode, $this->client->response->status());
+
+        if ($statusCode == 500)
+            return;
+
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $widgetOptions = (new XiboHls($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($nameNew, $widgetOptions->name);
-        $this->assertSame($durationNew, $widgetOptions->duration);
+
+        /** @var XiboHls $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboHls($this->getEntityProvider()))->hydrate($response[0]);
+
+        $this->assertSame($name, $checkWidget->name);
+        $this->assertSame($duration, $checkWidget->duration);
         
-        foreach ($widgetOptions->widgetOptions as $option) {
+        foreach ($checkWidget->widgetOptions as $option) {
             if ($option['option'] == 'uri') {
-                $this->assertSame($uriNew, urldecode(($option['value'])));
+                $this->assertSame($uri, urldecode(($option['value'])));
             }
         }
-    }
-
-    /**
-    * Delete
-    * @group broken
-    */
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create Google traffic widget
-        $hls = (new XiboHls($this->getEntityProvider()))->create('HLS stream', 1, 20, 'http://ceu.xibo.co.uk/hls/big_buck_bunny_adaptive_master.m3u8', 0, 0, $playlistId);
-        # Delete it
-        $this->client->delete('/playlist/widget/' . $hls->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
     }
 }
