@@ -1,7 +1,8 @@
 <?php
-/*
+/**
+ * Copyright (C) 2018 Xibo Signage Ltd
+ *
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2015-2018 Spring Signage Ltd
  *
  * This file is part of Xibo.
  *
@@ -21,91 +22,56 @@
 
 namespace Xibo\Tests\Integration\Widget;
 
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\OAuth2\Client\Entity\XiboText;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
 
+/**
+ * Class TextWidgetTest
+ * @package Xibo\Tests\Integration\Widget
+ */
 class TextWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/text/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
-        parent::tearDown();
-    }
-        /**
-     * @group add
-     * @dataProvider provideSuccessCases
-     */
-    public function testAdd($name, $duration, $useDuration, $effect, $speed, $backgroundColor, $marqueeInlineSelector, $text, $javaScript)
-    {
-        //parent::setupEnv();
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
 
-        $response = $this->client->post('/playlist/widget/text/' . $playlistId, [
-            'name' => $name,
-            'duration' => $duration,
-            'useDuration' => $useDuration,
-            'effect' => $effect,
-            'speed' => $speed,
-            'backgroundColor' => $backgroundColor,
-            'marqueeInlineSelector' => $marqueeInlineSelector,
-            'text' => $text,
-            'javaScript' => $javaScript
-            ]);
-        $this->assertSame(200, $this->client->response->status());
-        $this->assertNotEmpty($this->client->response->body());
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $textOptions = (new XiboText($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($name, $textOptions->name);
-        $this->assertSame($duration, $textOptions->duration);
-        foreach ($textOptions->widgetOptions as $option) {
-            if ($option['option'] == 'effect') {
-                $this->assertSame($effect, $option['value']);
-            }
-            if ($option['option'] == 'speed') {
-                $this->assertSame($speed, intval($option['value']));
-            }
-            if ($option['option'] == 'text') {
-                $this->assertSame($text, $option['value']);
-            }
-        }
+        parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
 
     /**
@@ -115,7 +81,7 @@ class TextWidgetTest extends LocalWebTestCase
      */
     public function provideSuccessCases()
     {
-        # Sets of data used in testAdd
+        // Sets of data used in testAdd
         return [
             'text 1' => ['Text item', 10, 1, 'marqueeRight', 5, null, null, 'TEST API TEXT', null],
             'text with formatting' => ['Text item 2', 20, 1, 'marqueeLeft', 3, null, null, '<p><span style=color:#FFFFFF;><span style=font-size:48px;>TEST</span></span></p>', null],
@@ -123,55 +89,40 @@ class TextWidgetTest extends LocalWebTestCase
         ];
     }
 
-    public function testEdit()
+    /**
+     * Test Edit
+     * @dataProvider provideSuccessCases
+     */
+    public function testEdit($name, $duration, $useDuration, $effect, $speed, $backgroundColor, $marqueeInlineSelector, $text, $javaScript)
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a text widget with wrapper
-        $text = (new XiboText($this->getEntityProvider()))->create('Text item', 10, 1, 'marqueeRight', 5, null, null, 'TEST API TEXT', null, $playlistId);
-        $nameNew = 'Edited Name';
-        $durationNew = 80;
-        $textNew = 'Edited Text';
-        $response = $this->client->put('/playlist/widget/' . $text->widgetId, [
-            'name' => $nameNew,
-            'duration' => $durationNew,
-            'useDuration' => 1,
-            'effect' => $text->effect,
-            'speed' => $text->speed,
-            'backgroundColor' => null,
-            'marqueeInlineSelector' => null,
-            'text' => $textNew,
-            'javaScript' => null
+        $this->client->put('/playlist/widget/' . $this->widgetId, [
+            'name' => $name,
+            'duration' => $duration,
+            'useDuration' => $useDuration,
+            'effect' => $effect,
+            'speed' => $speed,
+            'backgroundColor' => $backgroundColor,
+            'marqueeInlineSelector' => $marqueeInlineSelector,
+            'text' => $text,
+            'javaScript' => $javaScript
             ], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+
         $this->assertSame(200, $this->client->response->status());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $textOptions = (new XiboText($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($nameNew, $textOptions->name);
-        $this->assertSame($durationNew, $textOptions->duration);
-                foreach ($textOptions->widgetOptions as $option) {
+
+        /** @var XiboText $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboText($this->getEntityProvider()))->hydrate($response[0]);
+
+        $this->assertSame($name, $checkWidget->name);
+        $this->assertSame($duration, $checkWidget->duration);
+
+        foreach ($checkWidget->widgetOptions as $option) {
             if ($option['option'] == 'text') {
-                $this->assertSame($textNew, $option['value']);
+                $this->assertSame($text, $option['value']);
             }
         }
-    }
-
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a text widget with wrapper
-        $text = (new XiboText($this->getEntityProvider()))->create('Text item', 10, 1, 'marqueeRight', 5, null, null, 'TEST API TEXT', null, $playlistId);
-        # Delete it
-        $this->client->delete('/playlist/widget/' . $text->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
     }
 }
