@@ -21,11 +21,8 @@
 
 namespace Xibo\Tests\Integration\Widget;
 
-use Xibo\Entity\DataSet;
-use Xibo\Entity\Layout;
-use Xibo\OAuth2\Client\Entity\XiboDataSet;
+use Xibo\Entity\Display;
 use Xibo\OAuth2\Client\Entity\XiboDisplay;
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\OAuth2\Client\Entity\XiboSchedule;
 use Xibo\OAuth2\Client\Entity\XiboTicker;
 use Xibo\OAuth2\Client\Exception\XiboApiException;
@@ -41,17 +38,16 @@ use Xibo\Tests\LocalWebTestCase;
 class TickerWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
-
     use DisplayHelperTrait;
 
-    /** @var Layout[] */
-	private $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
 
-	/** @var DataSet[] */
-	private $startDataSets;
+    /** @var int */
+    protected $widgetId;
 
-	/** @var XiboDisplay */
-	private $display;
+    /** @var XiboDisplay */
+    protected $display;
 
     // <editor-fold desc="Init">
 
@@ -74,10 +70,42 @@ class TickerWidgetTest extends LocalWebTestCase
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        $this->startDataSets = (new XiboDataSet($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/ticker/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
+
+        // Create a Display
+        $this->display = $this->createDisplay();
+        $this->displaySetLicensed($this->display);
+
+        // Schedule the Layout "always" onto our display
+        //  deleting the layout will remove this at the end
+        $event = (new XiboSchedule($this->getEntityProvider()))->createEventLayout(
+            date('Y-m-d H:i:s', time()+3600),
+            date('Y-m-d H:i:s', time()+7200),
+            $this->publishedLayout->campaignId,
+            [$this->display->displayGroupId],
+            0,
+            NULL,
+            NULL,
+            NULL,
+            0,
+            0
+        );
+
+        $this->displaySetStatus($this->display, Display::$STATUS_DONE);
     }
 
     /**
@@ -85,105 +113,15 @@ class TickerWidgetTest extends LocalWebTestCase
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    $this->getLogger()->error('Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
-        // tearDown all datasets that weren't there initially
-        $finalDataSets = (new XiboDataSet($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-
-        $difference = array_udiff($finalDataSets, $this->startDataSets, function ($a, $b) {
-            /** @var XiboDataSet $a */
-            /** @var XiboDataSet $b */
-            return $a->dataSetId - $b->dataSetId;
-        });
-
-        # Loop over any remaining datasets and nuke them
-        foreach ($difference as $dataSet) {
-            /** @var XiboDataSet $dataSet */
-            try {
-                $dataSet->deleteWData();
-            } catch (\Exception $e) {
-                $this->getLogger()->error('Unable to delete ' . $dataSet->dataSetId . '. E: ' . $e->getMessage());
-            }
-        }
-
-        if ($this->display !== null) {
-            // Delete the resources we've created (deleting the display also clears out the event)
-            $this->deleteDisplay($this->display);
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
 
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
 
     //</editor-fold>
-
-    /**
-     * Each array is a test run
-     * Format ($sourceId, $uri, $duration, $useDuration)
-     * @return array
-     */
-    public function provideAddSuccessCases()
-    {
-        # Sets of data used in testAdd
-        return [
-            'Feed' => [true, 'http://localhost/rss/feed.xml', 70, 1],
-            'DataSet' => [false, null, 80, 1]
-        ];
-    }
-
-    /**
-     * @dataProvider provideAddSuccessCases
-     */
-    public function testAdd($isFeed, $uri, $duration, $useDuration)
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-        
-        if ($isFeed) {
-            $this->client->post('/playlist/widget/ticker/' . $playlistId, [
-                'uri' => $uri,
-                'duration' => $duration,
-                'useDuration' => $useDuration,
-                'sourceId' => 1
-            ]);
-        } else {
-            # Create a new dataset
-            $dataSet = (new XiboDataSet($this->getEntityProvider()))->create('phpunit dataset', 'phpunit description');
-
-            $this->client->post('/playlist/widget/ticker/' . $playlistId, [
-                'dataSetId' => $dataSet->dataSetId,
-                'duration' => $duration,
-                'useDuration' => $useDuration,
-                'sourceId' => 2
-            ]);
-        }
-
-        // Check response
-        $this->assertSame(200, $this->client->response->status());
-
-        // Get the Ticker we created back out.
-        // TODO: why are we getting this out by its Playlist ID?!
-        $widgetOptions = (new XiboTicker($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($duration, $widgetOptions->duration);
-    }
 
     /**
      * @return array
@@ -202,7 +140,7 @@ class TickerWidgetTest extends LocalWebTestCase
                     'updateInterval' => 100,
                     'effect' => 'fade',
                     'speed' => 5,
-                    'copyright' => 'Copyrights ©Spring Signage',
+                    'copyright' => 'Copyrights ©Xibo Signage',
                     'itemsSideBySide' => 0,
                     'upperLimit' => 0,
                     'lowerLimit' => 0,
@@ -298,7 +236,7 @@ class TickerWidgetTest extends LocalWebTestCase
                     'updateInterval' => 100,
                     'effect' => 'fade',
                     'speed' => 5,
-                    'copyright' => 'Copyrights ©Spring Signage',
+                    'copyright' => 'Copyrights ©Xibo Signage',
                     'itemsSideBySide' => 0,
                     'upperLimit' => 0,
                     'lowerLimit' => 0,
@@ -323,27 +261,24 @@ class TickerWidgetTest extends LocalWebTestCase
      */
     public function testEditFeed($newWidgetOptions, $expectedDuration)
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        $this->getLogger()->debug('Create ticker for ' . $playlistId);
-
-        # Create a ticker with wrapper
-        $ticker = (new XiboTicker($this->getEntityProvider()))->create(1, 'http://xibo.org.uk/feed', null, 70, 1, $playlistId);
+        $this->getLogger()->debug('testEditFeed - IN');
 
         // Edit ticker widget
-        $this->client->put('/playlist/widget/' . $ticker->widgetId, $newWidgetOptions, ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+        $this->client->put('/playlist/widget/' . $this->widgetId, $newWidgetOptions, ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
 
-        $this->assertSame(200, $this->client->response->status());
+        $this->getLogger()->debug('Check Response');
+
+        // Check response
+        $this->assertSame(200, $this->client->response->status(), $this->client->response->getBody());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
 
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
 
         // Get the edited ticker back out again.
-        $editedTicker = (new XiboTicker($this->getEntityProvider()))->getById($playlistId);
+        /** @var XiboTicker $editedTicker */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $editedTicker = (new XiboTicker($this->getEntityProvider()))->hydrate($response[0]);
 
         // check if changes were correctly saved
         $this->assertSame('Edited widget', $editedTicker->name);
@@ -377,24 +312,11 @@ class TickerWidgetTest extends LocalWebTestCase
             }
         }
 
-        // Create a Display and Schedule my Layout to it
-        $this->display = $this->createDisplay();
-        $this->displaySetLicensed($this->display);
+        $this->getLogger()->debug('Test Complete - publish and validate');
 
-        // Schedule the Layout "always" onto our display
-        //  deleting the layout will remove this at the end
-        $event = (new XiboSchedule($this->getEntityProvider()))->createEventLayout(
-            date('Y-m-d H:i:s', time()+3600),
-            date('Y-m-d H:i:s', time()+7200),
-            $layout->campaignId,
-            [$this->display->displayGroupId],
-            0,
-            NULL,
-            NULL,
-            NULL,
-            0,
-            0
-        );
+        // Publish
+        $this->publishedLayout = $this->publish($this->publishedLayout);
+        $layout = $this->publishedLayout;
 
         // Confirm our Layout is in the Schedule
         $schedule = $this->getXmdsWrapper()->Schedule($this->display->license);
@@ -410,7 +332,14 @@ class TickerWidgetTest extends LocalWebTestCase
         // Parse the output, check the "duration" field.
         $this->getLogger()->debug('Calling GetResource - for ' . $layout->layoutId . ' - ' . $layout->regions[0]->regionId . ' - ' . $editedTicker->widgetId);
 
-        $html = $this->getXmdsWrapper()->GetResource($this->display->license, $layout->layoutId, $layout->regions[0]->regionId, $editedTicker->widgetId);
+        $html = null;
+        try {
+            $html = $this->getXmdsWrapper()->GetResource($this->display->license, $layout->layoutId, $layout->regions[0]->regionId, $editedTicker->widgetId);
+            $this->getLogger()->debug('Get Resource Complete');
+        } catch (\Exception $exception) {
+            $this->getLogger()->error($html);
+            $this->fail($exception->getMessage());
+        }
 
         // Parse the HTML for the expected duration is per item
         $matches = [];
@@ -420,94 +349,7 @@ class TickerWidgetTest extends LocalWebTestCase
 
         $this->assertEquals(2, count($matches), 'More than 1 duration tag in HTML');
         $this->assertEquals($expectedDuration, intval($matches[1]), 'Duration doesn\'t match expected duration');
-    }
 
-     /**
-     * Edit dataSet ticker 
-     */
-    public function testEditDataset()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create dataset
-        $dataSet = (new XiboDataSet($this->getEntityProvider()))->create('phpunit dataset', 'phpunit description');
-        # Create a ticker with wrapper
-        $ticker = (new XiboTicker($this->getEntityProvider()))->create(2, null, $dataSet->dataSetId, 30, 1, $playlistId);
-        # Edit ticker widget
-        $noDataMessage = 'no records found';
-        $response = $this->client->put('/playlist/widget/' . $ticker->widgetId, [
-                'sourceId' => 2,
-                'dataSetId' => $dataSet->dataSetId,
-                'name' => 'Edited widget',
-                'duration' => 90,
-                'useDuration' => 1,
-                'updateInterval' => 100,
-                'effect' => 'fadeout',
-                'speed' => 500,
-                'template' => '[Col1]',
-                'durationIsPerItem' => 1,
-                'itemsSideBySide' => 1,
-                'upperLimit' => 0,
-                'lowerLimit' => 0,
-                'itemsPerPage' => 5,
-                'noDataMessage' => $noDataMessage
-        ], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
-        $this->assertSame(200, $this->client->response->status());
-        $this->assertNotEmpty($this->client->response->body());
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $widgetOptions = (new XiboTicker($this->getEntityProvider()))->getById($playlistId);
-        # check if changes were correctly saved
-        $this->assertSame('Edited widget', $widgetOptions->name);
-        $this->assertSame(90, $widgetOptions->duration);
-        foreach ($widgetOptions->widgetOptions as $option) {
-            if ($option['option'] == 'updateInterval') {
-                $this->assertSame(100, intval($option['value']));
-            }
-            if ($option['option'] == 'effect') {
-                $this->assertSame('fadeout', $option['value']);
-            }
-            if ($option['option'] == 'speed') {
-                $this->assertSame(500, intval($option['value']));
-            }
-            if ($option['option'] == 'template') {
-                $this->assertSame('[Col1]', $option['value']);
-            }
-            if ($option['option'] == 'durationIsPerItem') {
-                $this->assertSame(1, intval($option['value']));
-            }
-            if ($option['option'] == 'itemsSideBySide') {
-                $this->assertSame(1, intval($option['value']));
-            }
-            if ($option['option'] == 'upperLimit') {
-                $this->assertSame(0, intval($option['value']));
-            }
-            if ($option['option'] == 'lowerLimit') {
-                $this->assertSame(0, intval($option['value']));
-            }
-            if ($option['option'] == 'itemsPerPage') {
-                $this->assertSame(5, intval($option['value']));
-            }
-            if ($option['option'] == 'noDataMessage') {
-                $this->assertSame($noDataMessage, $option['value']);
-            }
-        }
-    }
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a ticker with wrapper
-        $ticker = (new XiboTicker($this->getEntityProvider()))->create(1, 'http://localhost/rss/feed.xml', null, 70, 1, $playlistId);
-        # Delete it
-        $this->client->delete('/playlist/widget/' . $ticker->widgetId);
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
+        $this->getLogger()->debug('Test Complete');
     }
 }
