@@ -20,7 +20,6 @@
  */
 namespace Xibo\Tests\Integration\Widget;
 
-use Xibo\OAuth2\Client\Entity\XiboLayout;
 use Xibo\OAuth2\Client\Entity\XiboWebpage;
 use Xibo\Tests\Helper\LayoutHelperTrait;
 use Xibo\Tests\LocalWebTestCase;
@@ -29,82 +28,45 @@ class WebpageWidgetTest extends LocalWebTestCase
 {
     use LayoutHelperTrait;
 
-	protected $startLayouts;
+    /** @var \Xibo\OAuth2\Client\Entity\XiboLayout */
+    protected $publishedLayout;
+
+    /** @var int */
+    protected $widgetId;
+
     /**
      * setUp - called before every test automatically
      */
     public function setup()
-    {  
+    {
         parent::setup();
-        $this->startLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
+
+        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+
+        // Create a Layout
+        $this->publishedLayout = $this->createLayout();
+
+        // Checkout
+        $layout = $this->checkout($this->publishedLayout);
+
+        // Create a Widget for us to edit.
+        $response = $this->getEntityProvider()->post('/playlist/widget/webpage/' . $layout->regions[0]->regionPlaylist['playlistId']);
+
+        $this->widgetId = $response['widgetId'];
     }
+
     /**
      * tearDown - called after every test automatically
      */
     public function tearDown()
     {
-        // tearDown all layouts that weren't there initially
-        $finalLayouts = (new XiboLayout($this->getEntityProvider()))->get(['start' => 0, 'length' => 10000]);
-        # Loop over any remaining layouts and nuke them
-        foreach ($finalLayouts as $layout) {
-            /** @var XiboLayout $layout */
-            $flag = true;
-            foreach ($this->startLayouts as $startLayout) {
-               if ($startLayout->layoutId == $layout->layoutId) {
-                   $flag = false;
-               }
-            }
-            if ($flag) {
-                try {
-                    $layout->delete();
-                } catch (\Exception $e) {
-                    fwrite(STDERR, 'Unable to delete ' . $layout->layoutId . '. E:' . $e->getMessage());
-                }
-            }
-        }
+        // Delete the Layout we've been working with
+        $this->deleteLayout($this->publishedLayout);
+
         parent::tearDown();
+
+        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
     }
-
-	/**
-	 * @group add
-     * @dataProvider provideSuccessCases
-     */
-	public function testAdd($name, $duration, $useDuration, $transparency, $uri, $scaling, $offsetLeft, $offsetTop, $pageWidth, $pageHeight, $modeId)
-	{
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-		$response = $this->client->post('/playlist/widget/webpage/' . $playlistId, [
-        	'name' => $name,
-        	'duration' => $duration,
-            'useDuration' => $useDuration,
-        	'transparency' => $transparency,
-        	'uri' => $uri,
-        	'scaling' => $scaling,
-        	'offsetLeft' => $offsetLeft,
-        	'offsetTop' => $offsetTop,
-        	'pageWidth' => $pageWidth,
-        	'pageHeight' => $pageHeight,
-        	'modeId' => $modeId
-        	]);
-        $this->assertSame(200, $this->client->response->status());
-        $this->assertNotEmpty($this->client->response->body());
-        $object = json_decode($this->client->response->body());
-        $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $webpageOptions = (new XiboWebpage($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($name, $webpageOptions->name);
-        $this->assertSame($duration, $webpageOptions->duration);
-        foreach ($webpageOptions->widgetOptions as $option) {
-            if ($option['option'] == 'uri') {
-                $this->assertSame($uri, urldecode($option['value']));
-            }
-            if ($option['option'] == 'modeId') {
-                $this->assertSame($modeId, intval($option['value']));
-            }
-        }
-	}
 
 	/**
      * Each array is a test run
@@ -123,62 +85,45 @@ class WebpageWidgetTest extends LocalWebTestCase
         ];
     }
 
-    public function testEdit()
+    /**
+     * @throws \Xibo\OAuth2\Client\Exception\XiboApiException
+     * @dataProvider provideSuccessCases
+     */
+    public function testEdit($name, $duration, $useDuration, $transparency, $uri, $scaling, $offsetLeft, $offsetTop, $pageWidth, $pageHeight, $modeId)
     {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a webpage with wrapper
-    	$webpage = (new XiboWebpage($this->getEntityProvider()))->create('Open natively webpage widget', 60, 1, NULL, 'http://xibo.org.uk/', NULL, NULL, NULL, NULL, NULL, 1, $playlistId);
-    	$nameNew = 'Edited Name';
-    	$durationNew = 80;
-    	$modeIdNew = 3;
-    	$uriNew = 'http://xibo.org.uk/about/';
-    	$response = $this->client->put('/playlist/widget/' . $webpage->widgetId, [
-        	'name' => $nameNew,
-        	'duration' => $durationNew,
-            'useDuration' => 1,
-        	'transparency' => $webpage->transparency,
-        	'uri' => $uriNew,
-        	'scaling' => $webpage->scaling,
-        	'offsetLeft' => $webpage->offsetLeft,
-        	'offsetTop' => $webpage->offsetTop,
-        	'pageWidth' => $webpage->pageWidth,
-        	'pageHeight' => $webpage->pageHeight,
-        	'modeId' => $modeIdNew
+        $response = $this->client->put('/playlist/widget/' . $this->widgetId, [
+        	'name' => $name,
+        	'duration' => $duration,
+            'useDuration' => $useDuration,
+        	'transparency' => $transparency,
+        	'uri' => $uri,
+        	'scaling' => $scaling,
+        	'offsetLeft' => $offsetLeft,
+        	'offsetTop' => $offsetTop,
+        	'pageWidth' => $pageWidth,
+        	'pageHeight' => $pageHeight,
+        	'modeId' => $modeId
         	], ['CONTENT_TYPE' => 'application/x-www-form-urlencoded']);
+
         $this->assertSame(200, $this->client->response->status());
         $this->assertNotEmpty($this->client->response->body());
         $object = json_decode($this->client->response->body());
         $this->assertObjectHasAttribute('data', $object, $this->client->response->body());
-        $webpageOptions = (new XiboWebpage($this->getEntityProvider()))->getById($playlistId);
-        $this->assertSame($nameNew, $webpageOptions->name);
-        $this->assertSame($durationNew, $webpageOptions->duration);
-        foreach ($webpageOptions->widgetOptions as $option) {
+
+        /** @var XiboWebpage $checkWidget */
+        $response = $this->getEntityProvider()->get('/playlist/widget', ['widgetId' => $this->widgetId]);
+        $checkWidget = (new XiboWebpage($this->getEntityProvider()))->hydrate($response[0]);
+
+        $this->assertSame($name, $checkWidget->name);
+        $this->assertSame($duration, $checkWidget->duration);
+
+        foreach ($checkWidget->widgetOptions as $option) {
             if ($option['option'] == 'uri') {
-                $this->assertSame($uriNew, urldecode($option['value']));
+                $this->assertSame($uri, urldecode($option['value']));
             }
             if ($option['option'] == 'modeId') {
-                $this->assertSame($modeIdNew, intval($option['value']));
+                $this->assertSame($modeId, intval($option['value']));
             }
         }
-    }
-
-    public function testDelete()
-    {
-        // Create layout
-        $layout = $this->createLayout();
-        $layout = $this->checkout($layout);
-        $playlistId = $layout->regions[0]->regionPlaylist['playlistId'];
-
-        # Create a clock with wrapper
-		$webpage = (new XiboWebpage($this->getEntityProvider()))->create('Open natively webpage widget', 60, 1, NULL, 'http://xibo.org.uk/', NULL, NULL, NULL, NULL, NULL, 1, $playlistId);
-		# Delete it
-		$this->client->delete('/playlist/widget/' . $webpage->widgetId);
-        # This should return 204 for success
-        $response = json_decode($this->client->response->body());
-        $this->assertSame(200, $response->status, $this->client->response->body());
     }
 }
